Thread down the `purs` option (#47)
We want to make sure we use whatever `purs` was given.

We use the same `purs` option that the rest of the program uses. This
should prevent the issue where someone doesn't have `purs` available on
the path, but knows the location of a `purs` they want to use.

It also prevents a hard to diagnose error where there is a `purs` binary
available on the path, someone wants to use a different `purs` binary,
yet we're deciding which version to use based on what the `purs` on the
path said it was.